backend/ravi/requirements.txt
anyio==3.7.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
Babel==2.10.3
bcrypt==3.2.2
beautifulsoup4==4.12.2
beniget==0.4.1
blinker==1.6.2
branca==0.7.2
Brlapi==0.8.4
Brotli==1.0.9
certifi==2023.7.22
chardet==5.1.0
cliapp==1.20180812.1
click==8.1.6
cmdtest==0.32+git
colorama==0.4.6
contourpy==1.0.7
cryptography==38.0.4
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
distlib==0.3.7
distro==1.8.0
distro-info==1.5+ubuntu0.23.10.1
Django==4.2.4
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
duplicity==1.2.2
fasteners==0.17.3
fastjsonschema==2.18.0
filelock==3.12.2
folium==0.16.0
fonttools==4.38.0
freetype-py==2.4.0
fs==2.4.16
fuse-python==1.0.5
future==0.18.2
gast==0.5.2
GDAL==3.7.1
gpg==1.18.0
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httplib2==0.20.4
httpx==0.23.3
idna==3.3
importlib-metadata==4.12.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.4
jsonpointer==2.0
jsonschema==4.10.3
jupyter_core==5.3.1
keyring==24.2.0
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
libvirt-python==9.0.0
lockfile==0.12.2
louis==3.26.0
lxml==4.9.3
lz4==4.0.2+dfsg
Mako==1.2.4.dev0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.3
mdurl==0.1.2
monotonic==1.6
more-itertools==10.1.0
mpmath==0.0.0
nala==0.14.0
nbformat==5.9.1
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
OWSLib==0.29.2
packaging==23.1
paramiko==2.12.0
pbr==5.11.1
pexpect==4.8.0
Pillow==10.0.0
pipenv==2023.7.23
platformdirs==3.10.0
plotly==5.15.0
ply==3.11
psycopg2==2.9.6
ptyprocess==0.7.0
pycairo==1.24.0
pycups==2.0.1
Pygments==2.15.1
PyGObject==3.46.0
PyJWT==2.7.0
pylibacl==0.7.0
PyNaCl==1.5.0
pyparsing==3.1.0
pyproj==3.6.0
PyQt5==5.15.9
PyQt5-sip==12.12.2
pyrsistent==0.18.1
python-apt==2.6.0+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-decouple==3.8
pythran==0.11.0
pytz==2023.3
pyxattr==0.8.1
pyxdg==0.28
PyYAML==6.0.1
QScintilla==2.14.1
reportlab==4.0.4
requests==2.31.0
rfc3986==1.5.0
rich==13.3.1
rlPyCairo==0.3.0
scipy==1.10.1
SecretStorage==3.3.3
six==1.16.0
sniffio==1.2.0
socksio==1.0.0
soupsieve==2.4.1
sqlparse==0.4.2
ssh-import-id==5.11
sympy==1.12
systemd-python==235
tenacity==8.2.1
tomli==2.0.1
tornado==6.3.2
traitlets==5.5.0
ttystatus==0.38
typer==0.9.0
typing_extensions==4.7.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.14.0
ufw==0.36.2
unattended-upgrades==0.1
urllib3==1.26.16
usb-creator==0.3.16
virtualenv==20.24.3
virtualenv-clone==0.5.7
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
xyzservices==2024.4.0
zipp==1.0.0
